Create README.md
Looking for a collaboration on making a website for a small family business in South Africa. I am learning to code to able to run the website and marketing for the family business . I am not good with it , but has taken up the challenge so all help is appropriated.